requirements.txt
alembic==1.11.1
anyio==3.7.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.8
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
Werkzeug==2.3.4
